refactor: rename client.ts to becraft-client.ts for clarity
The file exports BeCraftClient (an API client class), so naming it
after the class makes the module's role unambiguous.
